migrasi 2

+++ form-pmb/database/migrations/2025_02_04_011239_create_i_ortu_table.php
@@ -28,7 +28,7 @@ return new class extends Migration
             $table->integer('penghasilan_ibu');
             $table->string('no_wa_ibu')->unique();
             $table->string('alamat_ortu');
-            $table->integer('iPribadi_id');
+            $table->unsignedBigInteger('iPribadi_id');
             $table->foreign('iPribadi_id')->references('id')->on('i_pribadi');
             $table->timestamps();
         });
